+++ src/main/java/com/lambdaschool/todos/repository/UsersRepository.java
@@ -3,7 +3,7 @@ package com.lambdaschool.todos.repository;
 import com.lambdaschool.todos.models.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepository extends JpaRepository<Users, Long>
+public interface UsersRepository extends JpaRepository<Users, Integer>
 {
     public Users findByUsername(String name);
 }
